perfect to lunch Online

+++ admin/includes/view_all_users.php
@@ -66,11 +66,16 @@
 
 <?php 
     if(isset($_GET['delete'])){
-        $the_user_id = $_GET['delete'];
-
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-    $delete_query = mysqli_query($connection, $query); 
-    header("Location: users.php");
+        if (isset($_SESSION['user_role'])) {
+            if ($_SESSION['user_role'] = 'admin') {
+                
+                $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+                
+                $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+                $delete_query = mysqli_query($connection, $query); 
+                header("Location: users.php");
+            }
+        } 
     }
 ?>
